requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.34.0
google-auth-oauthlib==1.2.2
gspread==6.1.4
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.1.2
numpy==2.3.2
oauthlib==3.3.1
packaging==24.2
pandas==2.3.1
pillow==10.4.0
plotly==5.24.0
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.27.0
rsa==4.9.1
scipy==1.16.1
six==1.17.0
smmap==5.0.2
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
